# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Regina caeli API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/patient/profile/update": {
            "patch": {
                "tags": [
                    "Auth"
                ],
                "operationId": "9ef0dfe79a7c071fab57d9909f031d8e",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "first_name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "last_name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "phone",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "gender",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "marital_status",
                        "in": "query",
                        "description": "marital_status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "religion",
                        "in": "query",
                        "description": "religion",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "preferred_language",
                        "in": "query",
                        "description": "religion",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nationality",
                        "in": "query",
                        "description": "nationality",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lga",
                        "in": "query",
                        "description": "lga",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "town",
                        "in": "query",
                        "description": "town",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "card_number",
                        "in": "query",
                        "description": "card_number",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "next_of_kin",
                        "in": "query",
                        "description": "next_of_kin",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "next_of_kin_phone",
                        "in": "query",
                        "description": "next_of_kin_phone",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nature_of_relationship",
                        "in": "query",
                        "description": "nature_of_relationship",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_of_birth",
                        "in": "query",
                        "description": "date_of_birth",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "insurance_number",
                        "in": "query",
                        "description": "insurance_number",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ward",
                        "in": "query",
                        "description": "ward",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state_of_residence",
                        "in": "query",
                        "description": "state_of_residence",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address_of_residence",
                        "in": "query",
                        "description": "address_of_residence",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/patient/get-users-created": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "operationId": "9894452a675fd505114b73057f356558",
                "responses": {
                    "200": {
                        "description": "Booking successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Code Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized Access",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Booking already exists",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}